# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Questkart API
  version: 1.0.0
  description: Internal API for the SSR platform. Contract-first API for SSR — a domain-agnostic tool that simulates a target
    audience with AI avatars (personas), has them react to candidate messages, scores those reactions by meaning (SSR), and
    ranks the messages to produce a recommendation + report.
servers:
- url: https://api.questkart.example/api/v1
  description: Production
- url: http://localhost:4010
  description: Prism mock server (contract-first development)
- url: http://localhost:8000/api/v1
  description: Local development
paths:
...
  /api/v1/avatars:
    get:
      summary: List avatars (personas)
      operationId: listAvatars
      responses:
        '200':
          description: A page of avatars.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarList'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: scope
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/AvatarScope'
          description: 'Filter by library vs study-scoped personas. (in: query)'
      - name: domain_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: query'
      - name: study_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: query'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          example: 20
          description: 'Max items per page. (in: query)'
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          description: 'Opaque cursor from a previous page’s next_cursor. (in: query)'
    post:
      summary: Create a (custom) avatar
      operationId: createAvatar
      responses:
        '201':
          description: Avatar created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        '422':
          description: Validation failed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarCreate'
  /api/v1/avatars/{avatar_id}:
    get:
      summary: Get an avatar
      operationId: getAvatar
      responses:
        '200':
          description: The avatar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        '404':
          description: Resource not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: avatar_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: path'
    patch:
      summary: Update an avatar
      operationId: updateAvatar
      responses:
        '200':
          description: Updated avatar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        '404':
          description: Resource not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: avatar_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: path'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarUpdate'
    delete:
      summary: Delete an avatar
      operationId: deleteAvatar
      responses:
        '204':
          description: Deleted.
        '404':
          description: Resource not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: avatar_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: path'
  /api/v1/studies/{study_id}/panel:
    get:
      summary: Get the study’s avatar panel
      operationId: getPanel
      responses:
        '200':
          description: The panel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Panel'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: study_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: path'
    put:
      summary: Replace the study’s avatar panel
      operationId: setPanel
      responses:
        '200':
          description: Updated panel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Panel'
        '422':
          description: Validation failed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: Unexpected error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: study_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          description: 'in: path'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PanelUpdate'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Bearer authentication using Azure AD. Clients obtain an Azure AD-issued JWT access token and pass
        it in the Authorization: Bearer <access_token> header for API authentication.'
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details.
      properties:
        type:
          type: string
          format: uri
          description: default about:blank
          example: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: No study exists with that id.
        instance:
          type: string
          format: uri
          example: https://api.questkart.example/v1/exports/…
        errors:
          type: array
          items:
            type: object
          description: Field-level validation errors (when applicable).
          example:
          - '{ … }'
    PageMeta:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Cursor for the next page, or null if none.
        has_more:
          type: boolean
          example: true
    Timestamps:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          description: read-only
          example: '2026-01-15T09:30:00Z'
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: read-only
          example: '2026-01-15T09:30:00Z'
          readOnly: true
...
    AvatarScope:
      type: string
      enum:
      - library
      - study
    AvatarSource:
      type: string
      enum:
      - prebuilt
      - custom
      - llm_assisted
    Avatar:
      type: object
      allOf:
      - $ref: '#/components/schemas/Timestamps'
      - type: object
        properties:
          created_at:
            type: string
            format: date-time
            description: read-only
            example: '2026-01-15T09:30:00Z'
            readOnly: true
          updated_at:
            type: string
            format: date-time
            description: read-only
            example: '2026-01-15T09:30:00Z'
            readOnly: true
          id:
            type: string
            format: uuid
            description: read-only
            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
            readOnly: true
          scope:
            $ref: '#/components/schemas/AvatarScope'
          domain_id:
            type: string
            format: uuid
            nullable: true
            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          study_id:
            type: string
            format: uuid
            nullable: true
            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          name:
            type: string
            example: The Evidence-Driven Specialist
          profile:
            type: string
            description: The full descriptive persona.
            example: Sub-specialist who weighs RCT data and guideline alignment
          source:
            $ref: '#/components/schemas/AvatarSource'
...
    AvatarCreate:
      type: object
      properties:
        name:
          type: string
          example: Q3 lead-claim test — PCPs
        profile:
          type: string
          example: Sub-specialist who weighs RCT data and guideline alignment
        scope:
          $ref: '#/components/schemas/AvatarScope'
        domain_id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        study_id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        source:
          $ref: '#/components/schemas/AvatarSource'
      required:
      - name
      - profile
    AvatarUpdate:
      type: object
      properties:
        name:
          type: string
          example: Q3 lead-claim test — PCPs
        profile:
          type: string
          example: Sub-specialist who weighs RCT data and guideline alignment
    AvatarList:
      type: object
      allOf:
      - $ref: '#/components/schemas/PageMeta'
      - type: object
        properties:
          next_cursor:
            type: string
            nullable: true
            description: Cursor for the next page, or null if none.
          has_more:
            type: boolean
            example: true
          data:
            type: array
            items:
              $ref: '#/components/schemas/Avatar'
    Panel:
      type: object
      properties:
        study_id:
          type: string
          format: uuid
          description: read-only
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          readOnly: true
        avatar_ids:
          type: array
          items:
            type: string
            format: uuid
          example:
          - 3fa85f64-5717-4562-b3fc-2c963f66afa6
    PanelUpdate:
      type: object
      properties:
        avatar_ids:
          type: array
          items:
            type: string
            format: uuid
          description: min items 1
          example:
          - 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
      - avatar_ids
...